Refectoring: Refector follow&seller&qna

+++ src/main/java/com/myomi/seller/dto/SellerOrderDetailDto.java
@@ -22,7 +22,7 @@ public class SellerOrderDetailDto {
 	@JsonIgnore
 	private Product product;
 	
-	private Long prodCnt;
+	private int prodCnt;
 	private Long prodNum;
 	private String name;
 	private Long orderNum;
@@ -41,7 +41,7 @@ public class SellerOrderDetailDto {
 	private LocalDateTime createdDate;
 
 	@Builder
-	public SellerOrderDetailDto(Order order, Product product, Long prodCnt, Long prodNum, String name, Long orderNum,
+	public SellerOrderDetailDto(Order order, Product product, int prodCnt, Long prodNum, String name, Long orderNum,
 			String user, int week, String msg, String deliveryName, String deliveryTel, String deliveryAddr,
 			String deliveryMsg, String receiveDate, LocalDateTime createdDate) {
 		this.order = order;
